+++ arelle/formula/XPathContext.py
@@ -49,7 +49,8 @@ from arelle.formula.XPathParser import (
     VariableRef,
     exceptionErrorIndication,
 )
-from arelle.XmlValidate import UNKNOWN, VALID, VALID_NO_CONTENT, validate as xmlValidate
+from arelle.XmlValidateConst import UNKNOWN, VALID, VALID_NO_CONTENT
+from arelle.XmlValidate import validate as xmlValidate
 from arelle.typing import TypeGetText
 
 if TYPE_CHECKING:
